css/cssom/xml-stylesheet-pi-in-doctype.xhtml

<!DOCTYPE html [<?xml-stylesheet href="data:text/css,html{z-index: 1}"?>]>
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>xml-stylesheet processing instruction in doctype internal subset</title>
    <link rel="help" href="https://www.w3.org/TR/2004/REC-xml-infoset-20040204/#infoitem.pi"/>
    <!-- "[parent] The document, element, or document type declaration information item
    which contains this information item in its [children] property. " -->
    <link rel="help" href="https://w3c.github.io/csswg-drafts/cssom/#prolog"/>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      promise_test(async () => {
        await new Promise(resolve => window.onload = resolve);
        assert_equals(getComputedStyle(document.documentElement).zIndex, "auto");
      });
    </script>
  </body>
</html>
