third_party/blink/web_tests/external/wpt/css/cssom/xml-stylesheet-pi-in-doctype-expected.txt

This is a testharness.js-based test.
[FAIL] xml-stylesheet processing instruction in doctype internal subset
  assert_equals: expected "auto" but got "1"
Harness: the test ran to completion.
